ld.cの分割途中のファイルです。コンパイルエラーが起きます 	modified:   ld.c 	new file:   rel.c 	new file:   sym.c 	new file:   sym.h 	new file:   util.c 	new file:   util.h
